backend/python/requirements.txt
numpy
pandas
scikit-learn
joblib
librosa
praat-parselmouth
